03_Advanced/10_Alternative_Agent_Frameworks/CrewAI/01_Foundations/Some_Simple_Agents/app/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
crewai[tools]
litellm
mem0ai
python-dotenv
requests
nest_asyncio
